LayoutTests/platform/wpe/imported/w3c/web-platform-tests/webstorage/storage_string_conversion.window-expected.txt


FAIL localStorage only stores strings The quota has been exceeded.
FAIL sessionStorage only stores strings assert_equals: expected "function(){}" but got "function (){}"
